actual bugfix for #793 (thanks doomedraven)
Thanks to doomedraven for pointing out that by replacing Dictionary with
a dict we would no longer be able to use attribute-style variables (not
sure if that's the name for such syntax, but talking about
self.task.category etc) which we use quite a lot in the Scheduler.

Fixing the actual issue - passing a non-dict to multiprocessing in
process.py - can therefore be fixed in a much simpler way (by explicitly
casting from Dictionary to dict).