update processing_get_task to be more compatible
Thanks to doskop and appeltje a somewhat better workaround has been
developed. By updating a row and then selecting it again we can ensure
that the current thread (or processing instance, in our case) actually
got the task assigned.

Furthermore, we utilize "FOR UPDATE" to improve row-locking on
databases that support this keyword. Apparently it's not possible for
MySQL/MariaDB to embed both the select and update query in one query,
hence we split it up (somewhat assuming that the locking remains
in-place).

Even though SQLite3 should not be used in this scenario, it should still
be possible with our current safeguards (although I'm assuming "FOR
UPDATE" is not a thing for SQLite3).