per-analysis monitoring mode, retract iexplore mode for now
We retract the "iexplore" monitoring mode from the iexplore package for
now as we really only support proper analysis for *one* version of
Internet Explorer 8. Any other version would likely result in close to
zero results, in the default mode.

In future updates we should add functionality to let users inform us
about the exact versions of Internet Explorer they are using so that we
can start supporting the most used ones.