give mongodb order 2
In legacy Cuckoo we accidentally didn't run into this issue, but what
is happening is that the MongoDB reporting module adjusts some fields
in the report. E.g., it stores the dump.pcap GridFS ObjectId into
"network.pcap_id". In turn the JSON reporting module is not able to
store this type of value and throws an exception.
By placing MongoDB as order two it will always be executed last (in
our current setup anyway).