two office static deobfuscation tweaks
oletools sometimes fails to decompress the VBA macro (I suspect this to
be a known bug as Office is in fact able to properly decompress it, to
be fixed upstream later).
For now don't turn Chr(int) into a string. An improvement would be to
only do this for ASCII characters, but in the case of obfuscated /
encrypted strings this causes Unicode issues in our deobfuscation
process.