tweak analyzer zipfile creation logic
Due to the OldGuestManager/GuestManager thing we had this piece of code
duplicated which isn't necessarily a good thing, so this commit merges
the two independent functions by a single function.
It also contains a performance optimization as StringIO.StringIO() has
terrible performance when - in this case - a lot of files are added to
the Zip file. Its replacement, io.BytesIO(), in some cases reduces the
speed of the Zip file generation from tens of seconds or even minutes
to at most one or two seconds.
Also gets rid of another usage of StringIO in the Guest Manager logic.