reintroduce opt-in datetime encoding for task objects
Due to a Cuckoo Core change [1] [2] the datetime fields of the Task
object would no longer be encoded according to our known datetime format
string ("%Y-%m-%d %H:%M:%S"). This results in the datetime object ending
up in flask.jsonify which in turn encodes the datetime into an
impossible format.
This commit returns the datetime as we knew it before, but only for the
Cuckoo API, just in case other parts of Cuckoo rely on this behavior.

[1]: https://github.com/cuckoosandbox/cuckoo/commit/b42a9a4ff80aab5fcf505c20137f138b00c0b3d3
[2]: https://github.com/cuckoosandbox/cuckoo/commit/0f744365892d43d6891f717519afc3a9adad9360