expose the git hash in the .cwd file
This Git hash will - as was already possible in legacy Cuckoo - help
identify the version of a users' installation when reporting issues as
well as help us identify if a CWD update is required upon updating the
Cuckoo installation.
I.e., there will be updates to the CWD that will have to be applied
gradually for users that don't setup a brand new Cuckoo.