move routing information to its own configuration file
Before we used to have a section in cuckoo.conf as well as the entire
vpn.conf file, both including partial configuration of the routing
information. We combine these two into a single routing.conf file,
improving the representation of the configuration now that we have added
the drop route, an InetSim route, and the Tor route as well.