emit auxiliary/processing/signature/reporting errors
We are continuously getting new exceptions here and there due to
ongoing developments and thus new functionality. Through this commit we
improve the way that such exceptions are logged, namely, not only to the
cuckoo.log file, but also to the database so that the report will
include said exceptions.
Through the Web Interfaces' feedback form users are able to submit these
issues to us (Core Developers) so that we can then resolve them.

Due to the order of execution, errors happening while running the
reporting modules may not be included, but everything up until that
should be fine.