I made the following changes to be able to successfully analyse android apps:
I have made changes to avd.py so that it does not create a copy of original AVD.
it runs the same AVD provided by refrence_machine var. this will allow us to
run the AVD from snapshot. I also disabled the headless mode because in my
experince AVD cannot be run from snapshot in headless mode.
the reason to run AVD from snapshot is that if you root an AVD and restart it,
superuser app failes to give superuser access to app(although the su binary
is in place and given the excutable permission). to fix this problem you have
to do the rooting process all over.but since running AVD from snapshot does
not go through boot process, we won't have that problem.

I have also made changes to lib/cuckoo/common/objects.py. I tried to analyse
some apk files. get_apk_entry was not able to find the package name and the
name of the MainActivity of most of the apps I tried to analyse.
I installed apktool and made get_apk_entry use this tool. apktool extracts
an apk file and I read the AndroidManifest.xml file. AndroidManifest.xml
contains the name of the package.this works for all apk files.

I also made changes to adb.py. I changed execute_sample function. this way
you don't need to know the name of the MainActivity.

other changes was made as a result of the changes I just described.