Add restart functionality for physical emulation
When conducting emulation with the physical machinery, the physical guest machine 
has to be rebooted prior to subsequent analyses. This is currently done by issuing an 
RPC to the Windows RPC server, which is sent via the SMB protocol. At times, there are 
some underlying issues in SMB that prevent the RPC to be executed in the intended manner, 
and as a result the machine doesn't reboot, leaving it in an unclean state for further analysis. 
This can be over come by adding the option of rebooting via Cuckoo agent's XML-RPC server
by registering the aforementioned functionality as a procedure on the server, and invoking it
in case the regular RPC request fails. An additional change is made to agent.py to accomplish this.