first draft on getting cuckoo ready to r/w configuration
This commit has been manually verified to result in pretty much the same
configuration as we'd expect before this commit took place. It affects a
bit of the abstract Machinery class and as such introduces unit testing
in that area.

Aside from moving all default values to cuckoo/common/config.py and
reusing the *.conf files as jinja2 templates this commit also introduces
the concept of having, e.g., comma-separated, lists in *.conf files.

The only todo item with regards to configuration checking is ensuring
that all referenced sections in these section lists actually exist.