LibVirt machinery should create the KVM domain
Currently, LibVirt machinery can determine the snapshot and revert
the VM state to it, but it jumps straight to waiting for a state
change without actually issuing a create call to the domain being
used.

Add a single line to call self.vms[label].create() after resolving
the snapshot state and prior to entering the wait state.