make sniffer work on windows in theory
Most of all this commit adds unit tests for the sniffer auxiliary
module. Do note that on Windows we'll probably have to use windump.exe
so we may want to ship that executable ourselves as well as
automatically fall back to that file in the configuration and use
different arguments, in the future.