future signature compatibility with older cuckoo versions
Just in case we ever get to a stage where our current Signature version
criteria is not good enough (e.g., support for multiple major releases)
we will be able to provide backwards compatibility through the enable()
method. With the new Signature inheritance support it will be possible
to backport certain features into the Cuckoo Community repository
(again, if this will ever be necessary).