by default replace the old cuckoo agent by the new one
The Cuckoo package is a great time to finally deprecate the older Cuckoo
Agent. By doing so, various issues reported fairly frequently will be
resolved in the future when users start deploying the new Cuckoo Agent,
e.g., the issue where xmlrpclib (the core RPC mechanism used by the old
Cuckoo Agent) doesn't handle larger files very well.
Note that the Cuckoo will remain backwards compatible with the old
Cuckoo Agent (so that VMs don't have to be updated) as well as that
VMCloak has already been pushing the new Cuckoo Agent for quite a while
(as it exposes a better API for doing various operations inside the VM).

As-is the repository for the new Cuckoo Agent, including some unit
tests, may be found at https://github.com/jbremer/agent.