Slight fix for the toggle control
- detail view toggle was not yet injected with descriptions
- removed a :before rule for this specific thing

note: the filetree css seems to make a lot of disturbances when new things are added to the detail view in the filetree selection list. I should check this out later. Probably some kind of cascading issues (detail view lives in the same root as the filetree HTML, I should improve its closure to improve the view css a bit)