unit test database migrations from cuckoo 1.1
So the database migration for tasks/tags table were missing from our 0.6
release towards the 1.1 release, this doesn't mean, however, that we
should still apply those migrations for later setups (i.e., 1.1 and
upwards). Also note that Cuckoo 1.1 was shipping the incorrect schema
version (the one for 0.6 instead of 1.1) so some shuffling around of
revision nummbers was required.