simplify and make feedback up-to-date
Also brings some changes into the way we make the POST request to the
Cuckoo Feedback backend. By default it will only report some of the
files that are present in an analysis (the ones we deem most useful).
Furthermore the POST request now has a "feedback" field with JSON
contents in it and a "file" field containing the analysis .zip file.

Conflicts:
	tests/test_feedback.py