Fixes
- textual change in the coffee screen (I think I'm going to rename this module :p)
- re-enabled a diff in the item options. It's now like this:

{
    changed_properties: ["network_routing","machine"],
    options: {
        "network_routing": "inetsim",
        "machine": "cuckoo2"
    }
}

I thought passing in the changed_properties array might come in handy for validating an object in the backend, if unwanted I can remove it!

that contains an array with all the changed properties on this object. you can check these against the options to only get the changed values.