simplify cuckoo-create file copying process
The shutil.copytree() function always invokes os.makedirs(), but in our
situation we already have the destination directory in-place. This is
mainly useful for deployments where the cuckoo user doesn't have the
correct permissions for making the $CWD directory in the first place
(i.e., root permissions were required to mkdir & chown it).