replace postsubmit page to take just the submit-id
Before the postsubmit page would take a list of task IDs in the URL (as
part of the GET request). Aside from the fact that this is rather ugly,
this also doesn't scale in the sense that URLs are limited to 2-4kb in
length. Fix this by adding a relationship between Task and Submit and
showing that relation in the postsubmit page.