remove some more obsolete files
And also port the allowZip64=True flag, a bugfix that would've otherwise
gone missing with code that got moved around. That said we should also
move to temporary files rather than the in-memory buffer that we're
using for exporting analyses right now. A thought on that; we'll need
some object that automatically deletes the file after the transfer,
perhaps a class that inherits "file"?