handle configparser-related configuration parsing bug
This is probably a feature on ConfigParser's side to support multi-line
values for certain keys, but it seems to have broken the Cuckoo setup
for at least a handful of people so far. So let's raise an accurate
exception early on to prevent those people from running into related
issues later on at runtime.