for now comment out the summary section
Currently nobody is using it and as long as it's just a way too long
list of strings, nobody is going to use it. Instead we should put more
effort into identifying interesting strings in the backend, but as one
can imagine, this takes a lot of time & patience. To be continued.