make distributed cuckoo work with mysql #646 #1197
Two small changes required to be compatible with MySQL, whose backend
doesn't like unique columns with an indefinitive length. Let's hope that
the PostgreSQL database migration gods are among us as we're not
providing a proper database migration for that setup (at this point
anyway), as I imagine that SQLAlchemy will work properly as-is.