move androguard requirement to a pip extra feature
This fully gets rid of the lxml dependency which appears to be a pain in
the ass on every operating system.
We should certainly be updating our documentation to reflect the various
extra's that one may install, depending on the setup, i.e.,
cuckoo[android] cuckoo[distributed] cuckoo[postgresql] etc.