include yara offsets and prepare for 3.5.0+ inclusion
In a couple of weeks there will be a new Yara release which we can
properly include as a dependency. With that release we'll be able to
bind Cuckoo to our version of choice, i.e., the latest version.
The new offset field will allow us to quickly locate where the strings
were found in the binary / memory / etc.
This probably breaks backwards compatibility with older versions of
Yara, but if no Yara release is around in time, we can just do a soft
error on versions that are not 3.5.0.
Also disables some kernel driver stuff that's not yet ready for
deployment.