handle files that are not in the archive root directory
Pretty important to handle this properly as, e.g., before the execution
would simply fail altogether due to invalid sample paths being given in
the VM. In the future we should start using JSON-based analysis options
as currently this functionality could be broken by embedding a comma in
a samples' filename. I believe there's already a start in this
direction (for our improved Android Analyzer), but some additional
unittesting should be performed naturally.