fixup non-functional screenshots and human auxiliary modules
Due to a recent commit related to the integration of zer0m0n [1] it has
been observed that the screenshots & human auxiliary modules aren't
functioning properly anymore.

The root cause for this is the fact that the Auxiliary class since [1]
has learned the "start" method, overriding threading.Thread's start()
routine. As such trying to invoke the two threads, which is done through
.start(), would result into calling a noop function rather than
threading's actual function.

This commit fixes that behavior by having threading.Thread take
precedence over the methods of Auxiliary (by reordering their
inheritance). Due to the multithreaded nature of this change, I'm not
sure if it's feasible to do unit tests on it. We'll have to wait for
some functional unit tests that should be implemented next time.

This commit is actually just two line changes, the rest is code cleanup.

[1]: 90eef30cd0fd6cb2236aeed96216d1ce743b8f8c