allow webif to be behind a https reverse proxy #1397 (thanks seanthegeek)
As per seanthegeek's comments these changes allow https to work without
harming the http handling as-is. Also moves the previously-optional
Django setting to the hardcoded settings part so that this option is no
longer optional and users with an older CWD don't require to upgrade
their CWD to apply the configuration changes.