additional volatility module loading tweaks and handling
Raises an exception for ImportError's not related to Volatility which
is useful in case a future update decides to add another dependency
that's not installed by default in our setups.
Removes the attempt at trying to unittest importing Volatility as this
is non-trivial in Python - it has been tested manually, though.