improved error reporting when mongodb is offline #1507 (thanks bcallaughan)
Due to MongoDB not connecting straight away when a MongoClient object is
created we didn't properly catch the issue of the MongoDB server being
offline. When running "cuckoo -d" this would cause a long stacktrace
(ending with a ServerSelectionTimeoutError exception) and "cuckoo web"
would start silently and simply fail later on when MongoDB is actually
required to fetch data.

From now on the MongoDB connection is tested upon startup avoiding these
issues (but introducing a slight additional delay).