Update _irma.html
When irma checks file with antiviruses and marks file as malisious we see table with results.
If file is clean we see result "No IRMA results available.". But if irma was terminated we see "No IRMA results available." too.
I delete string "and report.analysis.irma.status" to display table with antiviruses if probe results is clean.