suppress unrelated tcpdump not found warning
The following message is printed by scapy if tcpdump wasn't found in
$PATH. Note that, apparently, scapy invokes tcpdump at every "import
scapy" statement. In fact, it turns out that almost a dozen programs
are executed upon start (see also "strace -ffe execve cuckoo ...").

"WARNING: Failed to execute tcpdump. Check it is installed and in the PATH"

Thanks to dforce #1497 & alphastorm for reporting this issue.