cuckoo-web requires absolute cuckoo path #1424
Thanks to Cuckoo0000 and Waffles-2 for reporting this issue and
botherder for identifying the actual problem. This commit substitutes
the relative path in argv[0] by its absolute version before the
os.chdir() call, resolving the problem of not being able to locate the
relative filename after Django's HTTP server fork()/execv().

A more proper fix will have to be performed another time and will
include getting rid of all relative paths in the Cuckoo Web Interface.