community support for automated cwd migration
Ensures that doing a CWD migration for a CWD that has ran "cuckoo
community" works as advertised. Note that the older Community files,
that are backwards compatible with older Cuckoo versions, are
automatically patched now (we don't need to be backwards compatible here
anyway). Also adds version release comments to the various hashes.