Make rooter insert rules instead of adding them
By default, libvirt applies the following iptables rules to host-only
bridges on startup:

-P FORWARD DROP
-A FORWARD -i virbr0 -o virbr0 -j ACCEPT
-A FORWARD -o virbr0 -j REJECT --reject-with icmp-port-unreachable
-A FORWARD -i virbr0 -j REJECT --reject-with icmp-port-unreachable

Rooter must use the insert (-I) action rather than add (-A) so that the
new rules have a higher priority than existing ones. Otherwise,
forwarding and other commands will not work.

This is the root cause and fix for the routing issues I was having in
issue #1280.