pass along filename in nested archive submissions (thanks razuz)
So far we only unittested archives that are natively supported on each
OS. For those that require additional 3rd party tooling, such as .rar,
this caused issues with not being able to create the correct zip files.
Essentially the filename was passed along rather than the filepath,
causing the 3rd party tooling to look for a file that doesn't exist.