fix unable to open dump-sorted #1676 (thanks lazybluefrog)
In the event of an empty .pcap file (i.e., only the header is present)
our sort_pcap() method doesn't actually get to writing a new pcap
(provided dump_sorted.pcap is only created when there are packets
available). This commit fixes a related warning that'd be thrown in that
case. Also fixes a NoneType issue introduced [1].

[1]: db38a5093c766ce21a1b4ce73b2bd5a3d9f1c0ac