switch default recentfiles directory to my documents
In the past year people have been confused by the random files appearing
on the Desktop. With this change we modify this default path from the
Desktop to the "Documents" or "My Documents" directory (or a localized
version of that directory). This way it's less obvious (i.e., not to be
seen in the screenshots) and hopefully make less people confused.