remove errors.py and support file deletion in cwd migrations
It was wrong to use GetLastError() in this context anyway as we can't
ensure that CreateProcess() was the last kernel32 method invoked. Even
if it was, the last error wouldn't have made much sense anyway since
subprocess handles that just fine.
This commit also implements file removal for CWD migrations (and applies
it in this situation). This will also be applied / useful to Daan's
Linux Analyzer work where we're about to remove a couple of files.