do not recreate hashes.txt upon installation
There was an issue due to the hashes.txt file being recreated as
setup.py is being called upon installation. As we don't ship older
monitoring binaries these were then set to be deleted upon invocation of
Cuckoo. This is undesired as we'll want users to be easily able to use
older Cuckoo Monitor binaries.