Launching the web interface faster and in a safer way.
By default, django's web server forks itself to create another process that constantly scans the source code for changes, and than restarts the server.
This sometimes makes the launching of the web interface very slow + has a small potential to cause loss of state.
Passing a `--noreload` argument to django prevents this behavior.