upgrade to cuckoo agent 0.8 improving permission denied handling
In some cases, e.g., #1893 & #1876, it may happen that the Cuckoo Agent
fails to create a directory due to permission errors, effectively
breaking the GuestManager due to HTTP exceptions. Thanks to Daan Spitz,
this has now been resolved. This commit upgrades the Cuckoo Agent and
throws an error for instances with older version of the Cuckoo Agent.