perform static analysis before the extracted module
In the case of Office DDE files the Static module will run a number of
Yara rules and Extractor plugins. In order to record this in our report,
the Extracted processing module should be ran afterwards.

This commit also fixes a regression with regards to not recording PID
and first_seen information in the ExtractScripts class anymore.