revert mongodb sorting tweaks #2006
This was a good proposal by sebdg, however, it turns out that MongoDB
requires quite some (or at least, too much) memory for such sorting
rules, rendering this kind of useless on bigger'ish installations.

For what it's worth, these sorting rules were put in place so that,
after reprocessing an analysis, the latest results would be visible in
the Web Interface. Naturally the better and longterm solution will be to
completely delete any previous results. This is (still) a TODO item.

Thanks to seantree, nicpenning, and sebdg for identifying this issue and
providing a proper workaround. Keeping in mind the above the best way
forward seems to revert these changes for now, though.