work in progress on the new gevent-based resultserver
The "legacy" ResultServer consumes huge amounts of CPU & RAM resources
due to the way it operates, namely by creating a thread for each
incoming connection. Needless to say, in case of analyzing, e.g.,
Ransomware, with potentially thousands of dropped files this results
in thousands of connections and as such, Python threads.
The switch to Gevent will provide a huge performance boost for Cuckoo.