more accurately fetch the latest dist cuckoo task
When one manually submits a task to a Cuckoo node that's part of a
Distributed Cuckoo setup, then it's automatically picked up and removed
almost immediately upon completion by the Distributed Cuckoo Worker.

A more correct solution will be usage of the "owner" field (as per
several comments throughout the Distributed Cuckoo code), but for now
this is fine, too.