patch py27 subprocess to be unicode-aware
This commit introduces Unicode support to Python 2.7's process handling,
especially to the spawning of new processes which is handled by the
_subprocess module and uses CreateProcessA internally. Note that this
has already been resolved on Python 3's branch.

During our Windows unit tests on AppVeyor the lib/api/process.py is
imported during all unit tests, due to which all invocations of
_subprocess.CreateProcess go through our code. Therefore we handle some
edge cases that we'll never be needing in the Cuckoo Analyzer but are
nonetheless useful as reference implementation and/or for completeness.

Furthermore it should be noted that this changes the "trigger" attribute
in the configuration for Cuckoo Monitor from ASCII to UTF-8 which
requires [1] on the Cuckoo Monitor side to be used correctly. Due to
UTF-8 being ASCII compatible, this should be backwards compatible.

Finally it removes monitor/latest from hashes.txt and introduces a
cmd.exe-based Unicode command-line unit test as Python 2.7's python.exe
doesn't support Unicode command-line arguments either (apparently).

[1]: https://github.com/cuckoosandbox/monitor/commit/e071e63a66e831163a40abc45109fdf71fee829e