Debian 9 debugging symbol installation #2384
Hello,
I tried several time to install the debugging symbols with the provided (Line 45 to 58) without success.

So i used the provided package on debian linux-image-$(uname -r)-dbg and it works.

It's easier for user if they don't have to add another depot and key (for exemple keyserver.ubuntu.com is down currently).

/!\ But i don't know if using the standard image from the debina repo remove some functionality