delete vt query for dropped PE binaries
The vt result of dropped PE binaries stores row['virustotal'] in that code.
The valuable "row" is local valuable in for loop. 
Thus the it is not reduced to "results".

On the other hands, malware sometimes drops a lot of PE/MZ binaries.
It causes to waste the api license in many cases.