Pass a 'process' event when handling AndroidFileMonitor
The current way of handling behavioral events is that one 'process' event needs to be sent before handling any other events.
In case of Android, we have more than one behaviral log with separate handlers. So the best way to remedy this is by
emitting a process event for each handler.. that way the other handlers can recognize the process