linux/process: Support reading empty values from /proc/pid/status
The file /proc/<pid>/status provides information about a process in
the format 'key: value'. The parsing of this file failed if 'value'
was empty.

Add an dummy value that will be taken in case the real value from
the status file does not exist.

Additionally, update the log message from the parsing code to also
output the exception, if one is raised.

Closes: #3035